Move decoration handling functions to own file as pure functions
