+++ src/plugin/fieldViews/ProseMirrorFieldView.ts
@@ -3,13 +3,14 @@ import { DOMParser } from "prosemirror-model";
 import type { AttributeSpec, Mark, Node } from "prosemirror-model";
 import type { Plugin, Selection, Transaction } from "prosemirror-state";
 import { EditorState } from "prosemirror-state";
-import { Mapping, StepMap } from "prosemirror-transform";
+import { StepMap } from "prosemirror-transform";
 import type {
   DecorationGroup,
   DecorationSource,
   EditorProps,
 } from "prosemirror-view";
 import { DecorationSet, EditorView } from "prosemirror-view";
+import { getMappedDecorationsFromSource } from "../helpers/decorations";
 import type { PlaceholderOption } from "../helpers/placeholder";
 import { PME_UPDATE_PLACEHOLDER } from "../helpers/placeholder";
 import type { BaseFieldDescription } from "./FieldView";
@@ -354,42 +355,6 @@ export abstract class ProseMirrorFieldView extends FieldView<string> {
     this.decorationsPending = true;
   }
 
-  private getMappedDecorationsFromSet(
-    // A decoration set received, positioned relative to the outer editor. It may contain irrelevant decorations.
-    decorationSet: DecorationSet,
-    // The offset of the field from its containing element
-    fieldOffsetFromElement: number,
-    // The node representing the current field in the ProseMirror document.
-    fieldNode: Node
-  ) {
-    // This field may receive decorations that will not apply to its range.
-    // Find the decorations in the context of the original document that should apply to this field, based on the position of
-    // the decorations in the original field.
-    // We must filter the decorations before we offset them via a 'map', otherwise there may be errors due to decorations
-    // being mapped outside of a valid range.
-    const relevantDecosFromOriginalDoc = decorationSet.find(
-      fieldOffsetFromElement,
-      fieldOffsetFromElement + fieldNode.nodeSize
-    );
-    // We must recombine the Decoration[] into a DecorationSet because we can only 'map' a DecorationSet, not a Decoration[],
-    // and we want to reposition the decorations relative to the current field.
-    // We must do this in the context of the original document, because ProseMirror uses the structure of the document to structure
-    // the decorations - in particular Widget and Node decorations rely on the document structure for a correct DecorationSet to be
-    // formed.
-    const offsetMap = new Mapping([StepMap.offset(-fieldOffsetFromElement)]);
-    const relevantDecosInOriginalDoc = DecorationSet.create(
-      this.outerView.state.doc,
-      relevantDecosFromOriginalDoc
-    );
-    // We now 'map' the decorations based on the offset, so that they are positioned relative to the current field, rather
-    // than the outer doc.
-    const mappedDecorations = relevantDecosInOriginalDoc.map(
-      offsetMap,
-      fieldNode
-    );
-    return mappedDecorations;
-  }
-
   protected applyDecorationsFromOuterEditor(
     decorations: DecorationSource | DecorationSet | DecorationGroup,
     fieldNode: Node
@@ -406,35 +371,14 @@ export abstract class ProseMirrorFieldView extends FieldView<string> {
     // The node we will be altering is a child of its parent, so we add +1 to the offset.
     const fieldOffsetFromElement = this.offset + 1;
 
-    if ("find" in decorations) {
-      // 'decorations' is a DecorationSet. Map them, then set them as the field's decorations.
-      const relevantDecorationSet = this.getMappedDecorationsFromSet(
-        decorations,
-        fieldOffsetFromElement,
-        fieldNode
-      );
-      // Decorations may be lost if we don't recreate the DecorationSet in the context of the innerEditor
-      const decoArray = relevantDecorationSet.find();
-      const decosForField = DecorationSet.create(fieldNode, decoArray);
-      this.setDecorationsForEditor(decosForField);
-    } else if ("members" in decorations) {
-      // 'decorations' is a DecorationGroup. Map each member DecorationSet, combine them into a single DecorationSet,
-      // then set them as the field's decorations.
-      const relevantDecorations = decorations.members
-        .map((decorationSet) =>
-          this.getMappedDecorationsFromSet(
-            decorationSet,
-            fieldOffsetFromElement,
-            fieldNode
-          )
-        )
-        .flatMap((set) => set.find());
-      const decorationsAsSet = DecorationSet.create(
-        fieldNode,
-        relevantDecorations
-      );
-      this.setDecorationsForEditor(decorationsAsSet);
-    }
+    const mappedDecorationSet = getMappedDecorationsFromSource(
+      decorations,
+      fieldOffsetFromElement,
+      fieldNode,
+      this.outerView.state.doc
+    );
+
+    this.setDecorationsForEditor(mappedDecorationSet);
   }
 
   /**
